Make enum discriminants i64 instead of the host int.

+++ src/librustc/middle/trans/reflect.rs
@@ -273,7 +273,7 @@ pub impl Reflector {
             do self.bracketed(~"enum", extra) |this| {
                 for variants.eachi |i, v| {
                     let extra1 = ~[this.c_uint(i),
-                                   this.c_int(v.disr_val),
+                                   this.c_int(v.disr_val /*bad*/as int),
                                    this.c_uint(vec::len(v.args)),
                                    this.c_slice(
                                        bcx.ccx().sess.str_of(v.name))];
